+++ client-database-android/src/main/java/de/njsm/stocks/client/database/DatabaseModule.java
@@ -161,6 +161,11 @@ public interface DatabaseModule {
         return database.eventDao();
     }
 
+    @Provides
+    static PlotDao PlotDao(StocksDatabase database) {
+        return database.plotDao();
+    }
+
     @Binds
     ErrorRepository errorRepository(ErrorRepositoryImpl impl);