+++ client-fakes/src/main/java/de/njsm/stocks/client/business/FakeBusinessModule.java
@@ -479,4 +479,10 @@ public interface FakeBusinessModule {
     static TicketDisplayInteractor TicketDisplayInteractor() {
         return mock(TicketDisplayInteractor.class);
     }
+
+    @Provides
+    @Singleton
+    static FoodDetailsInteractor FoodDetailsInteractor() {
+        return mock(FoodDetailsInteractor.class);
+    }
 }